+++ src/variant.cpp
@@ -410,7 +410,7 @@ variant variant::operator%(const variant& v) const
 
 variant variant::operator^(const variant& v) const
 {
-	return variant(static_cast<int>(pow(as_int(), v.as_int())));
+	return variant(static_cast<int>(pow(static_cast<double>(as_int()), v.as_int())));
 }
 
 variant variant::operator-() const
